Monster Recall: Add Intelligent Tracking
The player shouldn't have to target to update the
currently tracked monster race. In fact, with the
auto_target option, one seldom targets at all. So,
we update the tracking automatically, provided the
monster list is being calculated. Only a valid
monster target in los will override the automatic
tracking.