Monster Recall: Add Intelligent Tracking
Fix up the logic for this. If there is no current track,
of if the current track is not in LoS, then auto update
to the first entry on the monster list. If there is a
target in LoS, then that really should already be tracked,
so double check that case.