Draconian Metamorphosis: Bug with Berserkers and weapon_ct
I'm guessing what happened to be the player hit CL35 in the
middle of a round of attacks. The call to mon_take_hit will
gain experience, check for gaining levels, let the user choose
their L35 power, then call handle_stuff to chuck all their
gear and call calc_bonuses. All before returning to the middle
of our now broken attack loop! But this is just a guess ...