Stop crashes with autotravel in the wilderness
If autotravel caused the map to e.g. scroll downwards while the target
location was in the top third of the map, it would have negative
coordinates and bad things would happen.

Autotravel will now avoid moving far enough away from the target to
scroll it off of the map.