Messages: Reduce the Spam
For a player with aggravation, the "It wakes up" messages are
irritating. Of course everything wakes up! Often, you get hundred
or so of these at once, especially after teleportation or
recall. We could check for player aggravation, but I decided to
bundle these up with the disturb_minor option instead.