Tweak replacementart object_level.
Previously replacementarts were a bit more powerful if you found them deeper
than the artifact level, which was strange. This is only a very minor nerf,
and as compensation we try increasing object_level if we are failing to reach
the desired power level (this is a minor buff to very powerful artifacts like
Vilya and the One Ring).