Windows Issues: Player Names and Repeat Last Command
isprint is a bit flaky on Windows. For an int out of the ASCII
range, it asserts under debug, but simply malfunctions under
release. This was causing the repeat last command to no longer
function.

Windows is not setting a player name on the 'New' option. Since
it is probably too much to rely on all ports to nicely do this,
py_birth now provides a default implementation.

I also fixed a couple of MSVC compiler warnings.