py_display: Fix display anomalies for player flags
I saw the problem on oook with a Leprechaun (by clouded) where
the flags info in the player column was all messed up. I couldn't
reproduce, but am pretty sure this will fix the issue. Asking
have_flag for -1 is weird and I assume we are just reading memory
out of bounds (which is why it doesn't repro).