Give "The X resists." messages for monsters that are only seen via telepathy.
This was very annoying - I would cast a lightning ball at something down a
dark corridor and assume from the lack of message that the monster didn't
resist electricity. There's no reason not to display these messages if the
player knows what the monster is, since it isn't an info leak.