Object Lore: Require Purchase to Learn from Shops
Hitherto, you could learn object flavors by simply entering a
shop and viewing its wares. Also, you could inspect ego items
to learn lore. By request, I am returning to the old requirement
that the player actually purchase the object to gain any lore.

Actually, the request was made with regard to flavors. However,
it really needs to apply to egos as well. If you obj_identify()
inside a shop (e.g. by examining an item), then the lore stuff
just happens, whether the object is a flavorful item or not. So,
to fix the issue for flavors, I had to make the restriction
apply to egos as well.