Chaos Vortex Playability
I did test the vortex race, but my evolution skipped the Chaos
vortex. It is indeed a bit harsh. (BTW, in designing the vortex,
I made polymorphing a monster no longer reset its health to full
and made sure it no longer aborts your round of attacks. If polymorphed,
you were supposed to still deal damage, but I missed the stupid line
to reset damage to zero ... sigh).

For this release, your melee attacks still do damage if the monster
is polymorphed. Also, I reduced the poly frequency by a factor of 5.
It should be OK now (you are supposed to be a bit weak until the end
anyway. Just watch out for chaos resistant foes).

Finally, oftentimes the polymorph would reselect the same race,
giving a meaningless message. I fixed this to omit the message.