Redo the 'g' (Get) Command
Historically, this was stay still a turn with flip pickup.
Most players won't know what that means and expect this command
to get objects under foot. Also, I'd like to address issues
with archery ammo and objects in general.

So 'g' is now a beefed up 'Get Objects' command. It gets any
objects under foot (autodestroying if destroy_get is set). Then
it initiates a travel sequence to get nearby autopicker objects
(if auto_get_objects is set) or nearby ammo (inscribed with =g if
auto_get_ammo is set). Upon completion of each stage in the travel
sequence, a new travel cmd is issued to the next nearest qualifying
object. The series terminates when there are no new objects; when
the player's pack becomes full; or when the player is disturbed
by a nearby monster, etc.

I tested this change quite a bit, but it's big, and may require
tweaking over time. Documentation is still pending.