Monster/Object Lists: Tab Toggles
Just like with the Inventory/Equipment sub-terminals. Personally,
I keep the Monster Recall on always, and toggle between the monster
list (usually on) and the object list (when no monsters are nearby).
I only have space for 2 child terminals, so being able to toggle
with a single key is useful.