Begin Large Rewrite
Things will be broken for quite some time. The goals here
are to implement a quiver. But the main obstacle here is
get_item which is an absolutely disgusting and unmanageable
mess. So, let's try to restore reason and sanity.

Plan of attack is to add new modules:
  [1] equip: manages player's equipped objects
  [2] pack: manages player's carried objects
  [3] quiver: manages the quiver.

These 3 modules will need to independently manage an inventory.
  [4] inv: helper module for the above 3. Could also
           be used by shop inventories.

Then we will replace get_item. Instead of returning an integer
with range -> location, we'll return a struct with all relevant
info, including the requested object.

Quivers will be a new tval. Player buys them, or finds them, and
then equips them (new slot). Arrows may be added to equipped
quivers only (by equipping the arrows). Shooting is from quiver
(or perhaps floor) only. Never the pack. Quivers have a max qty,
but allow for a large number of slots. Ego quivers. Quivers with
higher capacities. Etc. In short, I'll be busy for quite some time!