process_player now calls pack_lock
And pack_unlock, of course, and it does this before calling
process_command. The result: your pack can never resort in
the middle of a loop and pack indexes are *stable*. If some
bone-head calls notice_stuff() in the wrong place while your
code is looping, you need no longer feel nasty bugs :D

If a pack_optimize is missed, it is smart enough to reset
PN_OPTIMIZEPACK so that the normal notice_stuff in process_player
will get around to it.