Code Cleanup: obj_release
I simplified the client side a bit. Clients should always release
obj_prompt()'d objects. Interfacing with the autopicker, which may
or may not destroy the object, now delays the actual destruction
until obj_release(). This reduces client side errors in case the
autopicker actually does want to destroy the object.

The other problem area is wrt all the p_ptr->notice/redraw/update
stuff that I find tedious to rewrite everywhere. Historically, this
was overspecified (e.g. PW_INVEN when operating on equipment). I
was beginning to push this crap into obj_release(), but this was
a bad idea, so I factored out a couple of methods for clients to
use: gear_notice_id() and gear_notice_enchant(). There are situations
where obj_release() would be a kludge (e.g. enchant(), since the
object continues to be used *after* the release: a semantic error).